requirements.txt
psutil
pynput
Pillow
scapy
pyshark
requests
mitmproxy
pyxdg
pyautogui
python-xlib
python3-gi
python3-gi-cairo
gir1.2-gtk-3.0
evdev
websockets
rapidfuzz
